backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx

# MongoDB driver
pymongo[srv]

# Password hashing
passlib[bcrypt]

# JWT tokens
python-jose[cryptography]
requests